+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWHAAGQ0G1NKW966E9V9PWNN"
+        "01KWHDQ8V5JYTH5N94QPYBQGK6"
     )
     r"""The unique identifier of the data source"""