+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWHAAGPWQCA6R3ME5CZPDKJ1"
+        "01KWHDQ8V192XMVCX43BZGSZ7Y"
     )
     r"""The unique identifier of the data source"""