+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWHAAGQ3J1WKA8QZHAZQ6JKY"
+        "01KWHDQ8V7AFP4NJYSB241EDRH"
     )
     r"""The unique identifier of the data source"""