+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KWHAAGHYB08DVPHX6R027GSY"
+    id: Optional[str] = "01KWHDQ8Q23SH81M7N2TQ75AC7"
 
     description: Optional[str] = None
 
@@ -1357,7 +1357,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWHAAGHKH0D3979XSEM6RD1A"
+        "tool_01KWHDQ8PVW1R134B6W156MFBK"
     )
 
     display_name: Optional[str] = None
@@ -1488,7 +1488,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KWHAAGHHAEPRVQ8XK4PF30TW"
+    id: Optional[str] = "01KWHDQ8PTRQW32M1CDNW737WV"
 
     description: Optional[str] = None
 
@@ -1632,7 +1632,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWHAAGHG1AHYC5ZSE0E2D0PY"
+        "tool_01KWHDQ8PSE3PFBAKSDZ8AKM1J"
     )
 
     display_name: Optional[str] = None
@@ -1932,7 +1932,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWHAAGHC991A0XS47QRK2GBX"
+        "tool_01KWHDQ8PQP8XHTD13TTFBZDYG"
     )
 
     display_name: Optional[str] = None
@@ -2126,7 +2126,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWHAAGH9SV1MMTEZHPZYF6X6"
+        "tool_01KWHDQ8PN752J675ES23XY69Z"
     )
 
     display_name: Optional[str] = None
@@ -2324,7 +2324,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KWHAAGH7K4PJTF519ZVNCJ3E"
+        "tool_01KWHDQ8PKHHMM2KDCJMF2HF4Q"
     )
 
     display_name: Optional[str] = None