+++ packages/orq-rc/RELEASES.md
@@ -9209,3 +9209,13 @@ Based on:
 - [python v4.11.0-rc.71] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.71] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.71 - packages/orq-rc
+
+## 2026-07-02 12:46:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.787.0 (2.914.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.72] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.72] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.72 - packages/orq-rc